feat: 재고 관련 엔티티 추가 - 작물, 농약, 비료, 작물에 관한 재고 엔티티 추가

+++ src/main/java/com/cropkeeper/domain/cultivation/entity/Cultivation.java
@@ -28,5 +28,5 @@ public class Cultivation extends BaseTimeEntity {
     private CropVariety variety;
 
     @Column(name = "planting_amount")
-    private Integer plantingAmount;
+    private Long plantingAmount;
 }
